Treat the subdetid method correctly

+++ CalibCalorimetry/HcalPlugins/src/HcalHardcodeCalibrations.cc
@@ -47,8 +47,8 @@ namespace {
 
     if (result.empty()) {
       for (int eta = -HcalDetId::kHcalEtaMask2; 
-           eta <= HcalDetId::kHcalEtaMask2; eta++) {
-        for (int phi = 0; phi <= HcalDetId::kHcalPhiMask2; phi++) {
+           eta <= (int)(HcalDetId::kHcalEtaMask2); eta++) {
+        for (unsigned int phi = 0; phi <= HcalDetId::kHcalPhiMask2; phi++) {
           for (int depth = 1; depth < maxDepthHB + maxDepthHE; depth++) {
             for (int det = 1; det <= HcalForward; det++) {
 	      HcalDetId cell ((HcalSubdetector) det, eta, phi, depth);
